+++ examples/os-strings.rs
@@ -2,7 +2,7 @@ fn main() {
     println!("WhoAmI {}", env!("CARGO_PKG_VERSION"));
     println!();
     println!(
-        "User's Language        whoami::langs():               {:?}",
+        "User's Language        whoami::langs():                 {:?}",
         whoami::langs()
             .map(|l| {
                 l.map(|l| l.to_string()).collect::<Vec<String>>().join(", ")
@@ -10,36 +10,41 @@ fn main() {
             .unwrap_or_else(|_| "??".to_string()),
     );
     println!(
-        "User's Name            whoami::realname_os():         {:?}",
+        "User's Name            whoami::realname_os():           {:?}",
         whoami::realname_os(),
     );
     println!(
-        "User's Username        whoami::username_os():         {:?}",
+        "User's Username        whoami::username_os():           {:?}",
         whoami::username_os(),
     );
     println!(
-        "Device's Pretty Name   whoami::devicename_os():       {:?}",
+        "User's Account         whoami::fallible::account_os():  {:?}",
+        whoami::fallible::account_os()
+            .unwrap_or_else(|_| "<unknown>".to_string().into()),
+    );
+    println!(
+        "Device's Pretty Name   whoami::devicename_os():         {:?}",
         whoami::devicename_os(),
     );
     println!(
-        "Device's Hostname      whoami::fallible::hostname():  {:?}",
+        "Device's Hostname      whoami::fallible::hostname():    {:?}",
         whoami::fallible::hostname()
             .unwrap_or_else(|_| "localhost".to_string()),
     );
     println!(
-        "Device's Platform      whoami::platform():            {:?}",
+        "Device's Platform      whoami::platform():              {:?}",
         whoami::platform(),
     );
     println!(
-        "Device's OS Distro     whoami::distro():              {:?}",
+        "Device's OS Distro     whoami::distro():                {:?}",
         whoami::distro(),
     );
     println!(
-        "Device's Desktop Env.  whoami::desktop_env():         {:?}",
+        "Device's Desktop Env.  whoami::desktop_env():           {:?}",
         whoami::desktop_env(),
     );
     println!(
-        "Device's CPU Arch      whoami::arch():                {:?}",
+        "Device's CPU Arch      whoami::arch():                  {:?}",
         whoami::arch(),
     );
 }