+++ src/lib.rs
@@ -65,7 +65,7 @@
 )]
 #![doc(
     html_logo_url = "https://raw.githubusercontent.com/ardaku/whoami/v1/res/icon.svg",
-    html_favicon_url = "https://raw.githubusercontent.com/ardaku/whoami/v1/res/icon.svg",
+    html_favicon_url = "https://raw.githubusercontent.com/ardaku/whoami/v1/res/icon.svg"
 )]
 
 mod api;